rename launch file
