+++ tests/pytest_tests/unit_tests_old/tests_launch/test_launch_aws.py
@@ -5,7 +5,7 @@ from unittest.mock import MagicMock
 import pytest
 import wandb
 import wandb.sdk.launch._project_spec as _project_spec
-import wandb.sdk.launch.launch as launch
+import wandb.sdk.launch._launch as _launch
 from wandb.sdk.launch.environment.aws_environment import AwsEnvironment
 from wandb.sdk.launch.errors import LaunchError
 from wandb.sdk.launch.runner.sagemaker_runner import SagemakerSubmittedRun
@@ -109,7 +109,7 @@ def test_launch_aws_sagemaker_no_instance(
     kwargs["uri"] = uri
     kwargs["api"] = api
 
-    run = launch.launch(**kwargs)
+    run = _launch.launch(**kwargs)
     out, _ = capsys.readouterr()
     assert run.training_job_name == "test-job-1"
     assert "Project: test" in out
@@ -162,7 +162,7 @@ def test_launch_aws_sagemaker(
     kwargs = json.loads(fixture_open("launch/launch_sagemaker_config.json").read())
     kwargs["uri"] = uri
     kwargs["api"] = api
-    run = launch.launch(**kwargs)
+    run = _launch.launch(**kwargs)
     out, _ = capsys.readouterr()
     assert run.training_job_name == "test-job-1"
     assert "Project: test" in out
@@ -239,7 +239,7 @@ def test_launch_aws_sagemaker_launch_fail(
     kwargs["api"] = api
 
     with pytest.raises(LaunchError) as e_info:
-        launch.launch(**kwargs)
+        _launch.launch(**kwargs)
     assert "Failed to create training job when submitting to SageMaker" in str(
         e_info.value
     )
@@ -284,7 +284,7 @@ def test_sagemaker_specified_image(
     kwargs["resource_args"]["sagemaker"]["AlgorithmSpecification"][
         "TrainingInputMode"
     ] = "File"
-    launch.launch(**kwargs)
+    _launch.launch(**kwargs)
     out, _ = capsys.readouterr()
     assert "Project: test" in out
     assert "Entity: mock_server_entity" in out
@@ -389,7 +389,7 @@ def test_no_sagemaker_resource_args(
         kwargs["api"] = api
         kwargs["resource_args"].pop("sagemaker", None)
         with pytest.raises(LaunchError) as e_info:
-            launch.launch(**kwargs)
+            _launch.launch(**kwargs)
         assert (
             str(e_info.value)
             == "No sagemaker args specified. Specify sagemaker args in resource_args"
@@ -434,7 +434,7 @@ def test_no_OuputDataConfig(
         kwargs["api"] = api
         kwargs["resource_args"]["sagemaker"].pop("OutputDataConfig", None)
         with pytest.raises(LaunchError) as e_info:
-            launch.launch(**kwargs)
+            _launch.launch(**kwargs)
         assert (
             str(e_info.value)
             == "Sagemaker launcher requires an OutputDataConfig Sagemaker resource argument"
@@ -477,7 +477,7 @@ def test_no_StoppingCondition(
         kwargs["resource_args"]["sagemaker"].pop("StoppingCondition", None)
 
         with pytest.raises(LaunchError) as e_info:
-            launch.launch(**kwargs)
+            _launch.launch(**kwargs)
         assert (
             str(e_info.value)
             == "Sagemaker launcher requires a StoppingCondition Sagemaker resource argument"
@@ -517,7 +517,7 @@ def test_no_ResourceConfig(
         kwargs["resource_args"]["sagemaker"].pop("ResourceConfig", None)
 
         with pytest.raises(LaunchError) as e_info:
-            launch.launch(**kwargs)
+            _launch.launch(**kwargs)
         assert (
             str(e_info.value)
             == "Sagemaker launcher requires a ResourceConfig Sagemaker resource argument"
@@ -557,7 +557,7 @@ def test_no_RoleARN(
         kwargs["resource_args"]["sagemaker"].pop("RoleArn", None)
 
         with pytest.raises(LaunchError) as e_info:
-            launch.launch(**kwargs)
+            _launch.launch(**kwargs)
         assert (
             str(e_info.value)
             == "AWS sagemaker require a string RoleArn set this by adding a `RoleArn` key to the sagemaker"
