+++ tests/pytest_tests/unit_tests_old/tests_launch/test_launch_local_process.py
@@ -2,7 +2,7 @@ import sys
 
 import pytest
 import wandb
-import wandb.sdk.launch.launch as launch
+import wandb.sdk.launch._launch as _launch
 from wandb.sdk.launch.utils import PROJECT_SYNCHRONOUS
 
 from .test_launch import (
@@ -37,7 +37,7 @@ def test_launch_local_process_base_case(
         "project": "test",
         "resource": "local-process",
     }
-    run = launch.launch(**kwargs)
+    run = _launch.launch(**kwargs)
     run.wait()
     assert str(run.get_status()) == "finished"
     # # TODO: Why does this not work?
