+++ wandb/cli/cli.py
@@ -1281,7 +1281,7 @@ def launch(
     logger.info(
         f"=== Launch called with kwargs {locals()} CLI Version: {wandb.__version__}==="
     )
-    from wandb.sdk.launch import launch as wandb_launch
+    from wandb.sdk.launch._launch import launch
 
     api = _get_cling_api()
     wandb._sentry.configure_scope(process_context="launch_cli")
@@ -1334,7 +1334,7 @@ def launch(
     if queue is None:
         # direct launch
         try:
-            run = wandb_launch.launch(
+            run = launch(
                 api,
                 job,
                 entry_point,
