+++ wandb/sdk/launch/__init__.py
@@ -1,5 +1,5 @@
 from .agent.agent import LaunchAgent
-from .launch import launch
+from ._launch import launch
 from ._launch_add import launch_add
 
-__all__ = ["LaunchAgent", "launch", "launch_add"]
+__all__ = ["LaunchAgent", "_launch", "launch_add"]
